refactor: Update import paths for components and remove unused code
- Update import paths for components in UsersMapping.jsx, page.js, and Providers.jsx
- Remove unused QueryClientProvider and update import paths for components in Providers.jsx
- Remove unused mongoose.js and getCategory.js files
- Update import paths for components in page.js and display-validation-error.jsx
- Update import paths for components in utils/hooks/user/useGetAllUsers.js
- Update import paths for components in utils/actions/product/get-all-products.js
- Update import paths for components in utils/models/category.js
- Update import paths for components in app/(client)/produse/[id]/page.js and app/(client)/produse/page.js
- Update import paths for components in components/sections/products.jsx and components/layout/admin/admin-sidebar.jsx
- Update import paths for components in utils/actions/user/create-user.js
- Update import paths for components in app/(admin)/admin/utilizatori/page.js
- Update import paths for components in utils/functions/category/get-all-categories.js
- Update import paths for components in utils/functions/attribute/get-all-attributes.js
- Remove unused mongoose import in utils/mongoose.js
- Remove unused getCategory import in utils/actions/category/get-category.js
- Remove unused getProduct import in app/(client)/produse/[id]/page.js
- Remove unused getAllUsers import in app/(admin)/admin/utilizatori/page.js
- Remove unused getAllUsers import in utils/functions/user/get-all-users.js

+++ utils/actions/user/delete-user.js
@@ -24,6 +24,10 @@ export const deleteUser = authActionClient
   .action(async ({ parsedInput: { id } }) => {
     const user = await prisma.user.findUnique({ where: { id } });
 
+    if (!user) {
+      throw new Error("User not found");
+    }
+
     await prisma.user.delete({
       where: { id },
     });
